+++ src/js/src/views/Users.vue
@@ -197,7 +197,11 @@
       <div
         v-if="paginationNeeded"
         class="control is-flex is-align-items-center">
-        <b-switch v-model="table.isPaginated" size="is-small" type="is-light"
+        <b-switch
+          v-model="table.isPaginated"
+          @update:modelValue="changePaginate()"
+          size="is-small"
+          type="is-light"
           >Paginate</b-switch
         >
       </div>